+++ air/src/lib.rs
@@ -17,15 +17,6 @@ use miden_core::{
 pub mod config;
 mod constraints;
 
-mod errors;
-pub use errors::ExecutionOptionsError;
-
-mod options;
-pub use options::{DEFAULT_CORE_TRACE_FRAGMENT_SIZE, ExecutionOptions, ProvingOptions};
-
-mod proof;
-pub use proof::{ExecutionProof, HashFunction};
-
 pub mod trace;
 use trace::{AUX_TRACE_WIDTH, AuxTraceBuilder, MainTraceRow, TRACE_WIDTH};