+++ crates/lib/core/tests/crypto/falcon.rs
@@ -1,6 +1,6 @@
 use std::{sync::Arc, vec};
 
-use miden_air::{Felt, ProvingOptions};
+use miden_air::Felt;
 use miden_assembly::{Assembler, utils::Serializable};
 use miden_core::{EventName, ZERO, field::PrimeField64};
 use miden_core_lib::{CoreLibrary, dsa::falcon512_rpo};
@@ -8,6 +8,7 @@ use miden_processor::{
     AdviceInputs, AdviceMutation, DefaultHost, EventError, ExecutionError, OperationError,
     ProcessState, Program, ProgramInfo, StackInputs, crypto::RpoRandomCoin,
 };
+use miden_prover::ProvingOptions;
 use miden_utils_testing::{
     AdviceStackBuilder, Word,
     crypto::{
@@ -354,7 +355,7 @@ fn falcon_prove_verify() {
     host.register_handler(EVENT_FALCON_SIG_TO_STACK, Arc::new(push_falcon_signature))
         .unwrap();
 
-    let options = ProvingOptions::with_96_bit_security(miden_air::HashFunction::Blake3_256);
+    let options = ProvingOptions::with_96_bit_security(miden_core::HashFunction::Blake3_256);
     let (stack_outputs, proof) =
         prove_sync(&program, stack_inputs, advice_inputs, &mut host, options)
             .expect("failed to generate proof");