+++ processor/src/lib.rs
@@ -9,8 +9,13 @@ extern crate std;
 use alloc::vec::Vec;
 use core::fmt::{Display, LowerHex};
 
+pub use miden_air::trace::RowIndex;
 use miden_air::trace::{CHIPLETS_WIDTH, RANGE_CHECK_TRACE_WIDTH};
-pub use miden_air::{ExecutionOptions, ExecutionOptionsError, trace::RowIndex};
+
+mod execution_options;
+pub use execution_options::{
+    DEFAULT_CORE_TRACE_FRAGMENT_SIZE, ExecutionOptions, ExecutionOptionsError,
+};
 pub use miden_core::{
     AssemblyOp, EMPTY_WORD, Felt, Kernel, ONE, Operation, Program, ProgramInfo, StackInputs,
     StackOutputs, WORD_SIZE, Word, ZERO,